Partial fix for suspend/continue
This "works", except that after a continue and finishing the suspended input,
sometimes the *next* call to read-input() will soft hang on the first line,
checking if the input is a TTY (checking `$.input.t`).  Pressing <enter> will
un-hang it, and it seems to continue normally after that.

I'm not sure why this occurs, but I notice that the MoarVM implementation of
istty (https://github.com/MoarVM/MoarVM/blob/master/src/io/io.c#L43-L58) holds
a mutex while inspecting the handle.  No idea yet if this is the problem or
just a red herring.